LLM_Guard/requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
numpy>=1.26.4
pytest>=8.0.0